Move the settings components into a new settings package
