+++ app/src/main/AndroidManifest.xml
@@ -13,7 +13,7 @@
             android:supportsRtl="true"
             android:theme="@style/AppTheme">
         <activity
-                android:name=".SettingsActivity"
+                android:name=".settings.SettingsActivity"
                 android:label="@string/title_activity_settings">
         </activity>
