+++ app/src/main/java/com/example/hwutimetable/settings/SettingsActivity.kt
@@ -1,10 +1,11 @@
-package com.example.hwutimetable
+package com.example.hwutimetable.settings
 
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
 import androidx.fragment.app.DialogFragment
 import androidx.preference.Preference
 import androidx.preference.PreferenceFragmentCompat
+import com.example.hwutimetable.R
 
 class SettingsActivity : AppCompatActivity() {
 
@@ -13,7 +14,10 @@ class SettingsActivity : AppCompatActivity() {
         setContentView(R.layout.settings_activity)
         supportFragmentManager
             .beginTransaction()
-            .replace(R.id.settings, SettingsFragment())
+            .replace(
+                R.id.settings,
+                SettingsFragment()
+            )
             .commit()
         supportActionBar?.setDisplayHomeAsUpEnabled(true)
     }
